refactor: utils add size.ts
